+++ ui/packages/app/web/src/components/ui/Navbar.tsx
@@ -22,9 +22,8 @@ import {Link, LinkProps, useLocation, useNavigate} from 'react-router-dom';
 
 import {Button} from '@parca/components';
 import {Parca, ParcaSmall} from '@parca/icons';
-import {compareProfile} from '@parca/profile';
 import {selectDarkMode, useAppSelector} from '@parca/store';
-import {convertToQueryParams} from '@parca/utilities';
+import {convertToQueryParams, parseParams} from '@parca/utilities';
 
 import ReleaseNotesViewer from '../ReleaseNotesViewer';
 import ThemeToggle from './ThemeToggle';
@@ -59,6 +58,22 @@ const Navbar = () => {
   const queryParams = new URLSearchParams(location.search);
   const expressionA = queryParams.get('expression_a');
   const expressionB = queryParams.get('expression_b');
+  const comparing = queryParams.get('comparing');
+
+  const queryParamsURL = parseParams(window.location.search);
+
+  /* eslint-disable @typescript-eslint/naming-convention */
+  const {
+    from_a,
+    to_a,
+    merge_from_a,
+    merge_to_a,
+    time_selection_a,
+    filter_by_function,
+    dashboard_items,
+    selection_a,
+    expression_a,
+  } = queryParamsURL;
 
   const isComparePage = expressionA !== null && expressionB !== null;
 
@@ -73,17 +88,9 @@ const Navbar = () => {
   const compareExplanation =
     'Compare two profiles and see the relative difference between them more clearly.';
 
-  const isCurrentPage = (item: {label: string; href: string; external: boolean}) => {
-    if (item.href === 'compare' && isComparePage) {
-      return true;
-    }
-
-    if (!isComparePage && location.pathname === item.href) {
-      return true;
-    }
-
-    return false;
-  };
+  const isCurrentPage = (item: {label: string; href: string; external: boolean}) =>
+    (item.href === 'compare' && (isComparePage || comparing === 'true')) ||
+    (!isComparePage && comparing !== 'true' && location.pathname === item.href);
 
   const navigateTo = useCallback(
     (path: string, queryParams: any, options?: {replace?: boolean}) => {
@@ -97,6 +104,24 @@ const Navbar = () => {
     },
     [navigate]
   );
+
+  const queryToBePassed =
+    expression_a === undefined
+      ? {
+          comparing: 'true',
+        }
+      : {
+          comparing: 'true',
+          dashboard_items: dashboard_items,
+          expression_a: expression_a,
+          from_a: from_a,
+          to_a: to_a,
+          time_selection_a: time_selection_a,
+          selection_a: selection_a,
+          filter_by_function: filter_by_function,
+          merge_from_a: merge_from_a,
+          merge_to_a: merge_to_a,
+        };
 
   return (
     <Disclosure as="nav" className="relative z-10 dark:bg-gray-900">
@@ -164,7 +189,15 @@ const Navbar = () => {
                                   'rounded-none hover:no-underline border-b-2 focus:ring-0 focus:outline-none focus:ring-offset-0 h-full whitespace-nowrap font-medium'
                                 )}
                                 variant="link"
-                                onClick={() => compareProfile(navigateTo)}
+                                onClick={() =>
+                                  navigateTo(
+                                    '/',
+                                    {
+                                      ...queryToBePassed,
+                                    },
+                                    {replace: true}
+                                  )
+                                }
                                 onMouseEnter={() => setCompareHover(true)}
                                 onMouseLeave={() => setCompareHover(false)}
                                 id="h-compare-button"