+++ ui/packages/shared/utilities/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.78 (2024-06-27)
+
+**Note:** Version bump only for package @parca/utilities
+
 ## 0.0.77 (2024-06-05)
 
 **Note:** Version bump only for package @parca/utilities